remodeling/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.0.9
anyjson==0.3.3
appdirs==1.4.4
apturl==0.5.2
astroid==3.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
biopython==1.83
blinker==1.7.0
Brlapi==0.8.3
cached-property==1.5.2
cachetools==5.3.3
celery==5.2.3
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
dill==0.3.8
display==1.0.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ephem==4.1.3
fasteners==0.14.1
Flask==3.0.2
flatbuffers==24.3.7
flower==2.0.1
fonttools==4.50.0
frozenlist==1.4.1
future==0.18.2
gast==0.5.4
gff2json==0.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.20.2
humanize==4.9.0
idna==3.3
importlib-metadata==4.6.4
isort==5.13.2
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==3.2.0
keras==2.15.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
mymetal @ file:///home/rey/Documentos/dockermymetal/mymetal
namex==0.0.7
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
progress==1.6
prometheus_client==0.20.0
prompt-toolkit==3.0.28
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybiolib==1.1.1917
pycairo==1.20.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.1.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-memcached==1.58
python-version==0.0.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
rubber==1.6.0
scikit-learn==1.4.1.post1
scipy==1.12.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sutil==0.0.2
systemd-python==234
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
texttable==1.6.4
threadpoolctl==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
typed-ast==1.5.5
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==3.0.1
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==3.18.1
